Extract @tisyn/transport package from @tisyn/agent
Separates transport concerns (protocol session, remote agent middleware,
transport backends) from agent declarations (contracts, dispatch, invoke).

@tisyn/agent keeps: agent(), operation(), implementAgent(), Dispatch,
dispatch(), invoke(), and associated types.

@tisyn/transport gets: Transport interface, createSession(),
installRemoteAgent(), inprocessTransport(), and compliance suite.
