+++ packages/agent/src/index.ts
@@ -10,14 +10,3 @@ export type {
   ImplementationHandlers,
   Invocation,
 } from "./types.js";
-export { installRemoteAgent } from "./install-remote.js";
-export { inprocessTransport } from "./inprocess-transport.js";
-export type {
-  Transport,
-  AgentTransport,
-  AgentTransportFactory,
-  HostMessage,
-  AgentMessage,
-} from "./transport.js";
-export type { ProtocolSession, CreateSessionOptions } from "./session.js";
-export { createSession } from "./session.js";
