--- packages/agent/src/inprocess.test.ts
@@ -1,4 +1,0 @@
-import { inprocessTransport } from "./inprocess-transport.js";
-import { transportComplianceSuite } from "./transport-compliance.js";
-
-transportComplianceSuite("inprocess", inprocessTransport);
