+++ packages/transport/src/transports/inprocess.test.ts
@@ -1,0 +1,4 @@
+import { inprocessTransport } from "./inprocess.js";
+import { transportComplianceSuite } from "../transport-compliance.js";
+
+transportComplianceSuite("inprocess", inprocessTransport);
